scripts/requirements.txt
# PySpark and Delta Lake
pyspark==3.5.0
delta-spark==3.0.0

# Kafka
kafka-python==2.0.2


scikit-learn==1.4.2
mlflow==2.12.1


# Monitoring and HTTP
requests==2.31.0
prometheus-client==0.19.0


pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0

# Jupyter per analytics interattive
jupyter==1.0.0
jupyterlab==4.0.9

# Others
python-dotenv==1.0.0
psutil==5.9.6